% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbixSim.R
\name{createRandomRegressionDataset}
\alias{createRandomRegressionDataset}
\title{Create a random regression data set with binary class.}
\usage{
createRandomRegressionDataset(numRows, numCols)
}
\arguments{
\item{numRows}{\code{numeric} number of rows (samples).}

\item{numCols}{\code{numeric} number of columns (independent variables),}
}
\value{
\code{data.frame} with class column.
}
\description{
\code{createRandomRegressionDataset}
}
\examples{
ds <- createRandomRegressionDataset(100, 100)
}
\seealso{
Other simulation functions: \code{\link{createDiffCoexpMatrixNoME}},
  \code{\link{createDiffCoexpMatrixNull}},
  \code{\link{createDiffCoexpMatrix}},
  \code{\link{createMainEffectsMatrix}},
  \code{\link{createRandomMatrix}},
  \code{\link{getFoldChange}},
  \code{\link{randomNetworkSim}},
  \code{\link{scaleFreeSim}},
  \code{\link{simCorrMatrixNonUniform}},
  \code{\link{simCorrMatrix}}
}
\keyword{datagen}
