% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inbixGAIN.R
\name{dcgain}
\alias{dcgain}
\title{Differential coexpression genetic association interaction network algorithm,}
\usage{
dcgain(inbixData, verbose = FALSE)
}
\arguments{
\item{inbixData}{Data frame with samples in rows, genes in columns
and phenotype in the last column.
$param verbose Falg to send messages to stdout.}
}
\value{
results Matrix of gene by gene differential coexpression values.
}
\description{
\code{dcgain}
}
\examples{
data(testdata10)
rinbixDcgain <- dcgain(testdata10)
}
\seealso{
\code{\link{dmgain}} for differential modularity.
}

