% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inbixUtils.R
\name{fisherRtoZ}
\alias{fisherRtoZ}
\title{Fisher correlation transformation function R to Z distribution.}
\usage{
fisherRtoZ(x)
}
\arguments{
\item{x}{Correlation value -1 to 1.}
}
\value{
Transformed correlation value.
}
\description{
\code{fisherRtoZ}
}
\examples{
fisherRtoZ(seq(from=-1, to=1, by=0.25))
points(seq(from=-1, to=1, by=0.05), fisherRtoZ(seq(from=-1, to=1, by=0.05)))
}

