% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbixGAIN.R
\name{getInteractionEffects}
\alias{getInteractionEffects}
\title{Get interaction effects from generalized linear model regression.}
\usage{
getInteractionEffects(data, regressionFamily = "binomial",
  numCovariates = 0, writeBetas = FALSE, excludeMainEffects = FALSE,
  useBetas = FALSE, transformMethod = "", verbose = FALSE, numCores = 2)
}
\arguments{
\item{data}{Data frame with genes in columns and samples in rows.}

\item{regressionFamily}{String glm regression family name.}

\item{numCovariates}{Number of included covariates.}

\item{writeBetas}{Flag indicating whether to write beta values to separate file.}

\item{excludeMainEffects}{Flag indicating whether to exclude main effect terms.}

\item{useBetas}{Flag indicating betas rather than standardized betas used.}

\item{transformMethod}{String optional transform method.}

\item{verbose}{Flag to send verbose messages to stdout.}

\item{numCores}{Number of processor cores to use in mclapply}
}
\value{
results Matrix of gene by gene regression coefficients.
}
\description{
\code{getInteractionEffects}
}

