% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inbixRegression.R
\name{getMainEffect}
\alias{getMainEffect}
\title{Get the main effect coefficent of a variable using generalized linear regression - glm}
\usage{
getMainEffect(data, geneName, depVarName, regressionFamily)
}
\arguments{
\item{data}{Data frame with genes in columns and samples in rows.}

\item{geneName}{String column name of the gene used in the regression formula.}

\item{depVarName}{String phenotype column name used in the regression formula.}

\item{regressionFamily}{Stringregression family name for the glm.}
}
\value{
regression beta coefficient.
}
\description{
\code{getMainEffect}
}

