\name{modularityBestSplit}
\alias{modularityBestSplit}
\title{Find the best split of a modularity matrix using eigenvector decomposition}
\usage{
modularityBestSplit(B, m)
}
\arguments{
  \item{B}{modularity matrix}

  \item{m}{m}
}
\value{
list with modularity value Q and best split vector
}
\description{
\code{modularityBestSplit}
}

