% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbixFS.R
\name{rankUnivariateRegression}
\alias{rankUnivariateRegression}
\title{Rank variables by univariate regression.}
\usage{
rankUnivariateRegression(regressionData, numCores = 2, sortOrder = "pval")
}
\arguments{
\item{regressionData}{\code{data.frame} with variable in columns and samples in rows;
the last column should be labeled 'Class' and be 0 or 1 values.}

\item{numCores}{\code{numeric} number of processor cores to use in mclapply}

\item{sortOrder}{\code{character} sort retruned data frame by named column}
}
\value{
\code{data.frame} with variable, convergence status, beta coefficient, 
p-value, standard error and standardized beta columns.
}
\description{
\code{rankUnivariateRegression}
}
\examples{
data(testdata10)
rankUnivariateRegressionResults <- rankUnivariateRegression(testdata10)
}
\seealso{
Other feature selection functions: \code{\link{geneRank}},
  \code{\link{r2VIMorig}}, \code{\link{rankDcgainSnprank}},
  \code{\link{rankDeseq2}}, \code{\link{rankDeseq}},
  \code{\link{rankEdgeRGlm}}, \code{\link{rankEdgeR}},
  \code{\link{rankGeneRank}}, \code{\link{rankGlmnet}},
  \code{\link{rankIterativeRelieff}},
  \code{\link{rankLasso}}, \code{\link{rankLimma}},
  \code{\link{rankRandomForest}},
  \code{\link{rankRegainSnprank}},
  \code{\link{rankReliefSeq}}, \code{\link{rankRelieff}},
  \code{\link{rankSam}}, \code{\link{rankTTest}},
  \code{\link{snprankInbix}}, \code{\link{snprank}}
}
\keyword{models}
\keyword{regression}
\keyword{univar}
