% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inbixGAIN.R
\name{runInteractionEffectsTest}
\alias{runInteractionEffectsTest}
\title{Get the interaction effect of a pair of variables using generalized linear regression - glm.}
\usage{
runInteractionEffectsTest(data, variableIndices, depVarName, regressionFamily,
  numCovariates, excludeMainEffects)
}
\arguments{
\item{data}{Data frame with genes in columns and samples in rows.}

\item{variableIndices}{Column indices of variable pairs.}

\item{depVarName}{Name of the phenotype variable.}

\item{regressionFamily}{glm regression family name.}

\item{numCovariates}{Number of covariates included.}

\item{excludeMainEffects}{Flag indicating whether to exclude main effect terms.}
}
\value{
Data frame with gene, convergence status, beta coefficient, .
p-value, standard error and standardized beta columns.
}
\description{
\code{runInteractionEffectsTest}
}

