% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inbixFS.R
\name{snprank}
\alias{snprank}
\title{Rank genes by SNPrank algorithm.}
\usage{
snprank(G, gamma = 0.85)
}
\arguments{
\item{G}{Matrix genetic association interaction network.}

\item{gamma}{Numeric weighting, interactions (closer to 1) versus main effects (closer to 0).}
}
\value{
sortedTable Data frame with gene, SNPrank, diagonal and degree columns.
}
\description{
\code{snprank}
}
\examples{
data(testdata10)
rinbixRegain <- regainParallel(testdata10, stdBetas=TRUE, absBetas=TRUE)
rinbixSnpranksDF <- snprank(rinbixRegain)
}

