% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbixData.R
\docType{data}
\name{testdata10}
\alias{testdata10}
\title{A random dataset with 10 samples and 10 variables.}
\format{Data frame with 10 rows and 10 variables plus "Class" column.}
\usage{
data(testdata10)
}
\description{
A dataset containing random variables var1 ... var10. 
Class/phenotypes are 5 0's and 5 1's.
}
\keyword{datasets}
