% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inbixCppInterface.R
\name{writeRegressionDataAsInbixNumeric}
\alias{writeRegressionDataAsInbixNumeric}
\title{Write regression data frame as  inbix numeric data set and phenotype file.}
\usage{
writeRegressionDataAsInbixNumeric(regressionData, baseInbixName)
}
\arguments{
\item{regressionData}{Data frame with sample rows, gene columns
plus phenotype column.}

\item{baseInbixName}{Base filename to write.}
}
\description{
\code{writeRegressionDataAsInbixNumeric}
}

