\name{writeSimulatedInbixDataset}
\alias{writeSimulatedInbixDataset}
\title{Write data set as inbix format}
\usage{
writeSimulatedInbixDataset(D, Dfileprefix)
}
\arguments{
  \item{D}{is a data matrix}

  \item{Dfileprefix}{is an output data file prefix}
}
\description{
\code{writeSimulatedInbixDataset}
}

