% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createRandomRegressionDataset}
\alias{createRandomRegressionDataset}
\title{Create a random regression data set with binary class}
\usage{
createRandomRegressionDataset(numRows, numCols)
}
\arguments{
\item{numRows}{number of rows (samples)}

\item{numCols}{number of columns (independent variables)}
}
\value{
data frame
}
\description{
\code{createRandomRegressionDataset}
}

