% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inbixSim.R
\name{createRandomRegressionDataset}
\alias{createRandomRegressionDataset}
\title{Create a random regression data set with binary class.}
\usage{
createRandomRegressionDataset(numRows, numCols)
}
\arguments{
\item{numRows}{Numeric number of rows (samples).}

\item{numCols}{Numeric number of columns (independent variables),}
}
\value{
data frame with class column.
}
\description{
\code{createRandomRegressionDataset}
}

