% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbixGAIN.R
\name{dcgain}
\alias{dcgain}
\title{Differential coexpression genetic association interaction network (dcGAIN) algorithm.}
\usage{
dcgain(inbixData, verbose = FALSE)
}
\arguments{
\item{inbixData}{\code{data.frame} with samples in rows, genes in columns
and phenotype in the last column.}

\item{verbose}{\code{logical} to send messages to stdout.}
}
\value{
results \code{matrix} of gene by gene differential coexpression values.
}
\description{
\code{dcgain}
}
\examples{
data(testdata10)
rinbixDcgain <- dcgain(testdata10)
}
\references{
\itemize{
  \item \url{http://www.biodatamining.org/content/8/1/5}
  {Differential co-expression network centrality and machine learning feature selection for identifying susceptibility hubs in networks with scale-free structure}
  \item \url{https://github.com/hexhead/inbix}{C++ inbix on github}
}
}
\seealso{
\code{\link{dmgain}} for differential modularity. 
\code{\link{dcgainInbix}} for inbix differential coexpression GAIN.

Other GAIN functions: \code{\link{dcgainInbix}},
  \code{\link{dmgain}}, \code{\link{fitInteractionModel}},
  \code{\link{fitMainEffectModel}},
  \code{\link{gainToSimpleSIF}},
  \code{\link{getInteractionEffects}},
  \code{\link{getMainEffects}}, \code{\link{regainInbix}},
  \code{\link{regainParallel}}, \code{\link{regain}}

Other inbix synonym functions: \code{\link{geneLowValueFilter}},
  \code{\link{geneLowVarianceFilter}},
  \code{\link{getMainEffects}},
  \code{\link{modularityBestSplit}},
  \code{\link{modularity}}, \code{\link{regainParallel}},
  \code{\link{regain}}
}
\keyword{array}
\keyword{models}

