\name{dcgain}
\alias{dcgain}
\title{Differential coexpression genetic association network algorithm}
\usage{
dcgain(inbixData)
}
\arguments{
  \item{inbixData}{is a data frame with samples in rows,
  genes in columns and phenotype in the last column}
}
\value{
results a matrix of gene by gene differential coexpression
values
}
\description{
\code{dcgain}
}

