% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inbixGAIN.R
\name{dcgain}
\alias{dcgain}
\title{Differential coexpression genetic association interaction network algorithm,}
\usage{
dcgain(inbixData)
}
\arguments{
\item{inbixData}{Data frame with samples in rows, genes in columns
and phenotype in the last column.}
}
\value{
results Matrix of gene by gene differential coexpression values.
}
\description{
\code{dcgain}
}
\seealso{
\code{\link{dmgain}} for differential modularity.
}

