% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dmgain}
\alias{dmgain}
\title{Differential modularity genetic association network algorithm}
\usage{
dmgain(inbixData)
}
\arguments{
\item{inbixData}{is a data frame with samples in rows, genes in columns
and phenotype in the last column}
}
\value{
results a matrix of gene by gene differential modularity values
}
\description{
\code{dmgain}
}

