% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inbixGAIN.R
\name{dmgain}
\alias{dmgain}
\title{Differential modularity genetic association network algorithm.}
\usage{
dmgain(inbixData)
}
\arguments{
\item{inbixData}{Data frame with samples in rows, genes in columns
and phenotype in the last column.}
}
\value{
results Matrix of gene by gene differential modularity values.
}
\description{
\code{dmgain}
}

