% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbixUtils.R
\name{fisherRtoZ}
\alias{fisherRtoZ}
\title{Fisher correlation transformation function R to Z distribution.}
\usage{
fisherRtoZ(x)
}
\arguments{
\item{x}{\code{numeric} Correlation value -1 to 1.}
}
\value{
\code{numeric} Transformed correlation value.
}
\description{
\code{fisherRtoZ}
}
\examples{
fisherRtoZ(seq(from=-1, to=1, by=0.25))
}
\seealso{
Other utility functions: \code{\link{geneLowValueFilter}},
  \code{\link{geneLowVarianceFilter}},
  \code{\link{logSpiral}}, \code{\link{scaleAB}},
  \code{\link{sumOfPowers}}
}
\keyword{math}

