% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbixGAIN.R
\name{getMainEffects}
\alias{getMainEffects}
\title{Get main effects from generalized linear model regression (parallel).}
\usage{
getMainEffects(data, regressionFamily = "binomial", numCovariates = 0,
  writeBetas = FALSE, useBetas = FALSE, transformMethod = "",
  numCores = 2, verbose = FALSE)
}
\arguments{
\item{data}{\code{data.frame} with genes in columns and samples in rows.}

\item{regressionFamily}{\code{string} glm regression family name.}

\item{numCovariates}{\code{numeric}  of included covariates.}

\item{writeBetas}{\code{logical} indicating whther to write beta values to separate file.}

\item{useBetas}{\code{logical} indicating betas rather than standardized betas used.}

\item{transformMethod}{\code{string} optional transform method.}

\item{numCores}{\code{numeric} number of processor cores to use in mclapply}

\item{verbose}{\code{logical} to send verbose messages to stdout.}
}
\value{
mainEffectValues \code{vector} of main effect values.
}
\description{
\code{getMainEffects}
}
\seealso{
\code{\link{rankUnivariateRegression}}

Other GAIN functions: \code{\link{dcgainInbix}},
  \code{\link{dcgain}}, \code{\link{dmgain}},
  \code{\link{fitInteractionModel}},
  \code{\link{fitMainEffectModel}},
  \code{\link{gainToSimpleSIF}},
  \code{\link{getInteractionEffects}},
  \code{\link{regainInbix}}, \code{\link{regainParallel}},
  \code{\link{regain}}

Other inbix synonym functions: \code{\link{dcgain}},
  \code{\link{geneLowValueFilter}},
  \code{\link{geneLowVarianceFilter}},
  \code{\link{modularityBestSplit}},
  \code{\link{modularity}}, \code{\link{regainParallel}},
  \code{\link{regain}}
}
\keyword{array}
\keyword{internal}
\keyword{models}
\keyword{regression}
\keyword{univar}

