% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbixGAIN.R
\name{getMainEffects}
\alias{getMainEffects}
\title{Get main effects from generalized linear model regression (parallel).}
\usage{
getMainEffects(data, regressionFamily = "binomial", numCovariates = 0,
  writeBetas = FALSE, useBetas = FALSE, transformMethod = "",
  verbose = FALSE, numCores = 2)
}
\arguments{
\item{data}{Data frame with genes in columns and samples in rows.}

\item{regressionFamily}{String glm regression family name.}

\item{numCovariates}{Number of included covariates.}

\item{writeBetas}{Flag indicating whther to write beta values to separate file.}

\item{useBetas}{Flag indicating betas rather than standardized betas used.}

\item{transformMethod}{String optional transform method.}

\item{verbose}{Flag to send verbose messages to stdout.}

\item{numCores}{Number of processor cores to use in mclapply.}
}
\value{
mainEffectValues Vector of main effect values.
}
\description{
\code{getMainEffects}
}

