% Generated by roxygen2 (4.0.1): do not edit by hand
\name{modularity}
\alias{modularity}
\title{Detect modular structure in a network}
\usage{
modularity(G)
}
\arguments{
\item{G}{an adjacency matrix}
}
\value{
a data frame of genes name and module assignment columns
}
\description{
\code{modularity}
}

