% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbixCluster.R
\name{modularity}
\alias{modularity}
\title{Detect modular structure in a network.}
\usage{
modularity(G, verbose = FALSE)
}
\arguments{
\item{G}{\code{matrix} Adjacency matrix.}

\item{verbose}{\code{logical} to output messages to stdout.}
}
\value{
\code{list} with:
\itemize{
  \item \code{list} \code{groups} group assignments
  \item \code{numeric} \code{q} modularity
  \item \code{data.frame} \code{modules} modules
}
}
\description{
\code{modularity}
}
\examples{
data(testdata10)
rinbixRegain <- regainParallel(testdata10, stdBetas = TRUE, absBetas = TRUE)
rinbixModulesDF <- Rinbix::modularity(rinbixRegain)
}
\seealso{
Other cluster functions: \code{\link{loadRipmResultsFromRdata}},
  \code{\link{mergeSumPowers}},
  \code{\link{modularityBestSplit}},
  \code{\link{ripMKernelStack}}, \code{\link{ripM}},
  \code{\link{ripmVariableModuleAssignments}},
  \code{\link{saveModuleListGmt}},
  \code{\link{saveRipmModules}},
  \code{\link{saveRipmResultsToRdata}}

Other inbix synonym functions: \code{\link{dcgain}},
  \code{\link{geneLowValueFilter}},
  \code{\link{geneLowVarianceFilter}},
  \code{\link{getMainEffects}},
  \code{\link{modularityBestSplit}},
  \code{\link{regainParallel}}, \code{\link{regain}},
  \code{\link{snprank}}
}
\keyword{cluster}
\keyword{math}
