% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbixCluster.R
\name{modularity}
\alias{modularity}
\title{Detect modular structure in a network.}
\usage{
modularity(G, verbose = FALSE)
}
\arguments{
\item{G}{Matrix Adjacency matrix.}

\item{verbose}{Flag to output messages to stdout.}
}
\value{
list of groups, Q and module assignments.
}
\description{
\code{modularity}
}
\examples{
data(testdata10)
rinbixRegain <- regainParallel(testdata10, stdBetas=TRUE, absBetas=TRUE)
rinbixModulesDF <- Rinbix::modularity(rinbixRegain)
}

