% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inbixCluster.R
\name{modularity}
\alias{modularity}
\title{Detect modular structure in a network.}
\usage{
modularity(G)
}
\arguments{
\item{G}{Matrix Adjacency matrix.}
}
\value{
list of groups, Q and module assignments.
}
\description{
\code{modularity}
}

