% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inbixCluster.R
\name{modularityBestSplit}
\alias{modularityBestSplit}
\title{Find the best split of a modularity matrix using eigenvector decomposition.}
\usage{
modularityBestSplit(B, m)
}
\arguments{
\item{B}{Modularity matrix.}

\item{m}{m (average degree?)}
}
\value{
List with modularity value Q and best split vector.
}
\description{
\code{modularityBestSplit}
}

