% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbixCluster.R
\name{modularityBestSplit}
\alias{modularityBestSplit}
\title{Find the best split of a modularity matrix using eigenvector decomposition.}
\usage{
modularityBestSplit(B, m)
}
\arguments{
\item{B}{Matrix modularity matrix.}

\item{m}{Numeric average degree?}
}
\value{
List with modularity value Q and best split vector.
}
\description{
\code{modularityBestSplit}
}
\keyword{internal}

