% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbixFS.R
\name{rankUnivariateRegression}
\alias{rankUnivariateRegression}
\title{Rank genes by univariate regression.}
\usage{
rankUnivariateRegression(regressionData)
}
\arguments{
\item{regressionData}{\code{data.frame} with gene in columns and samples in rows;
the last column should be labeled 'Class' and be 0 or 1 values.}
}
\value{
\code{data.frame} with gene, convergence status, beta coefficient, 
p-value, standard error and standardized beta columns.
}
\description{
\code{rankUnivariateRegression}
}
\examples{
data(testdata100ME4)
rankUnivariateRegressionResults <- rankUnivariateRegression(testdata100ME4)
}
\seealso{
Other feature selection functions: \code{\link{geneRank}},
  \code{\link{r2VIMorig}}, \code{\link{rankDcgainSnprank}},
  \code{\link{rankDeseq2}}, \code{\link{rankDeseq}},
  \code{\link{rankEdgeRGlm}}, \code{\link{rankEdgeR}},
  \code{\link{rankGeneRank}}, \code{\link{rankGlmnet}},
  \code{\link{rankLasso}}, \code{\link{rankLimma}},
  \code{\link{rankRandomForest}},
  \code{\link{rankRegainSnprank}},
  \code{\link{rankReliefSeq}}, \code{\link{rankSam}},
  \code{\link{rankTTest}}, \code{\link{snprankInbix}},
  \code{\link{snprank}}
}
\keyword{models}
\keyword{regression}
\keyword{univar}

