% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbixFS.R
\name{rankUnivariateRegression}
\alias{rankUnivariateRegression}
\title{Rank genes by univariate regression.}
\usage{
rankUnivariateRegression(regressionData)
}
\arguments{
\item{regressionData}{Data frame with gene in columns and samples in rows;
the last column should be labeled 'Class' and be 0 or 1 values.}
}
\value{
Data frame with gene, convergence status, beta coefficient, 
p-value, standard error and standardized beta columns.
}
\description{
\code{rankUnivariateRegression}
}
\examples{
data(testdata100ME4)
rankUnivariateRegressionResults <- rankUnivariateRegression(testdata100ME4)
}

