% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inbixCppInterface.R
\name{readInbixNumericAsRegressionData}
\alias{readInbixNumericAsRegressionData}
\title{Read inbix numeric data set and phenotype file; return combined data frame.}
\usage{
readInbixNumericAsRegressionData(baseInbixName)
}
\arguments{
\item{baseInbixName}{Base filename.}
}
\value{
Data frame with numeric data in first m columns and phenotye in m+1 column.
}
\description{
\code{readInbixNumericAsRegressionData}
}

