% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbixCppInterface.R
\name{readInbixNumericAsRegressionData}
\alias{readInbixNumericAsRegressionData}
\title{Read inbix numeric data set and phenotype file; return combined data frame.}
\usage{
readInbixNumericAsRegressionData(baseInbixName)
}
\arguments{
\item{baseInbixName}{\code{string} base filename.}
}
\value{
\code{data.frame} with numeric data in first m columns and phenotye in m+1 column.
}
\description{
\code{readInbixNumericAsRegressionData}
}
\references{
\itemize{
  \item \url{https://github.com/hexhead/inbix}{C++ inbix on github}
}
}
\seealso{
Other inbix interface functions: \code{\link{dcgainInbix}},
  \code{\link{modularityInbix}},
  \code{\link{permuteGainInbix}},
  \code{\link{rankDcgainSnprank}},
  \code{\link{rankRegainSnprank}},
  \code{\link{regainInbix}}, \code{\link{snprankInbix}},
  \code{\link{writeRegressionDataAsInbixNumeric}}
}
