% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inbixGAIN.R
\name{regainParallel}
\alias{regainParallel}
\title{Parallel execution of the regression genetic association network algorithm.}
\usage{
regainParallel(regressionData, stdBetas = FALSE, absBetas = FALSE)
}
\arguments{
\item{regressionData}{Data frame with gene in columns and samples in rows;
the last column should be labeled 'Class' and be 0 or 1 values.}

\item{stdBetas}{Flag to use standardized beta coefficients.}

\item{absBetas}{Flag to use absolute value of beta coefficients.}
}
\value{
regainMatrix Matrix of gene by gene regression coefficients.
}
\description{
\code{regainParallel}
}

