% Generated by roxygen2 (4.0.1): do not edit by hand
\name{regainParallel}
\alias{regainParallel}
\title{Parallel execution of the regression genetic association network algorithm}
\usage{
regainParallel(regressionData, stdBetas = FALSE, absBetas = FALSE)
}
\arguments{
\item{regressionData}{a data frame with gene in columns and samples in rows;
the last column should be labeled 'Class' and be 0 or 1 values}

\item{stdBetas}{flag to use standardized beta coefficients}

\item{absBetas}{flag to use absolute value of beta coefficients}
}
\value{
regainMatrix a matrix of gene by gene regression coefficients
}
\description{
\code{regainParallel}
}

