% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbixGAIN.R
\name{runMainEffectsTest}
\alias{runMainEffectsTest}
\title{Get the main effect of a variable using generalized linear regression - glm.}
\usage{
runMainEffectsTest(data, variableName, depVarName, regressionFamily,
  numCovariates)
}
\arguments{
\item{data}{Data frame with genes in columns and samples in rows.}

\item{variableName}{String name of the variable to consider.}

\item{depVarName}{String name of the phenotype variable.}

\item{regressionFamily}{String glm regression family name.}

\item{numCovariates}{Number of included covariates.}
}
\value{
Data frame with gene, convergence status, beta coefficient, 
p-value, standard error and standardized beta columns.
}
\description{
\code{runMainEffectsTest}
}
\keyword{internal}

