% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inbixFS.R
\name{snprank}
\alias{snprank}
\title{Rank genes by SNPrank algorithm.}
\usage{
snprank(G, gamma = 0.85)
}
\arguments{
\item{G}{Genetic association network matrix.}

\item{gamma}{Parameter weighting interactions versus main effects.}
}
\value{
sortedTable Data frame with gene, SNPrank, diagonal and degree columns.
}
\description{
\code{snprank}
}

