% Generated by roxygen2 (4.0.2): do not edit by hand
\name{snprank}
\alias{snprank}
\title{Rank genes by SNPrank algorithm}
\usage{
snprank(G, gamma = 0.85)
}
\arguments{
\item{G}{is a genetic association network matrix}

\item{gamma}{is a parameter weighting interactions versus main effects}
}
\value{
sortedTable a data frame with gene, SNPrank, diagonal and degree columns
}
\description{
\code{snprank}
}

