% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbixGAIN.R
\name{snprank}
\alias{snprank}
\title{Rank variables by SNPrank algorithm.}
\usage{
snprank(G, alg.gamma = 0.85)
}
\arguments{
\item{G}{\code{matrix} genetic association interaction network.}

\item{alg.gamma}{\code{numeric} weighting, interactions (closer to 1) versus 
main effects (closer to 0).}
}
\value{
sortedTable \code{data.frame} with variable, SNPrank, diagonal and 
degree columns.
}
\description{
\code{snprank}
}
\examples{
data(testdata10)
rinbixRegain <- regainParallel(testdata10, stdBetas=TRUE, absBetas=TRUE)
rinbixSnpranksDF <- snprank(rinbixRegain)
}
\references{
\itemize{
  \item \url{http://www.nature.com/variable/journal/v11/n8/full/gene201037a.html}
  {Genes & Immunity Paper}
  \item \url{http://insilico.utulsa.edu/index.php/snprank/}{SNPrank Software}
}
}
\seealso{
Other feature selection functions: \code{\link{geneRank}},
  \code{\link{r2VIMorig}}, \code{\link{rankDcgainSnprank}},
  \code{\link{rankDeseq2}}, \code{\link{rankDeseq}},
  \code{\link{rankEdgeRGlm}}, \code{\link{rankEdgeR}},
  \code{\link{rankGeneRank}}, \code{\link{rankGlmnet}},
  \code{\link{rankIterativeRelieff}},
  \code{\link{rankLimma}}, \code{\link{rankRandomForest}},
  \code{\link{rankRegainSnprank}},
  \code{\link{rankReliefSeqInbix}},
  \code{\link{rankRelieff}}, \code{\link{rankSam}},
  \code{\link{rankTTest}},
  \code{\link{rankUnivariateRegression}},
  \code{\link{snprankInbix}}

Other GAIN functions: \code{\link{dcgainInbix}},
  \code{\link{dcgain}}, \code{\link{dmgain}},
  \code{\link{fitInteractionModel}},
  \code{\link{fitMainEffectModel}},
  \code{\link{gainToSimpleSIF}},
  \code{\link{getInteractionEffects}},
  \code{\link{getMainEffects}}, \code{\link{regainInbix}},
  \code{\link{regainParallel}}, \code{\link{regain}}

Other inbix synonym functions: \code{\link{dcgain}},
  \code{\link{geneLowValueFilter}},
  \code{\link{geneLowVarianceFilter}},
  \code{\link{getMainEffects}},
  \code{\link{modularityBestSplit}},
  \code{\link{modularity}}, \code{\link{regainParallel}},
  \code{\link{regain}}
}
