% Generated by roxygen2 (4.0.1): do not edit by hand
\name{writeRegressionDataAsInbixNumeric}
\alias{writeRegressionDataAsInbixNumeric}
\title{Write regression data frame as  inbix numeric data set and phenotype file}
\usage{
writeRegressionDataAsInbixNumeric(regressionData, baseInbixName)
}
\arguments{
\item{regressionData}{is a data frame with sample rows, gene columns
plus phenotype column}

\item{baseInbixName}{file base name}
}
\description{
\code{writeRegressionDataAsInbixNumeric}
}

