% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbixCppInterface.R
\name{writeRegressionDataAsInbixNumeric}
\alias{writeRegressionDataAsInbixNumeric}
\title{Write regression data frame as  inbix numeric data set and phenotype file.}
\usage{
writeRegressionDataAsInbixNumeric(regressionData, baseInbixName)
}
\arguments{
\item{regressionData}{\code{data.frame} with sample rows, gene columns 
plus phenotype column.}

\item{baseInbixName}{\code{string} base filename to write.}
}
\description{
\code{writeRegressionDataAsInbixNumeric}
}
\seealso{
Other inbix interface functions: \code{\link{dcgainInbix}},
  \code{\link{modularityInbix}},
  \code{\link{permuteGainInbix}},
  \code{\link{rankDcgainSnprank}},
  \code{\link{rankRegainSnprank}},
  \code{\link{readInbixNumericAsRegressionData}},
  \code{\link{regainInbix}}, \code{\link{snprankInbix}}
}

