% Generated by roxygen2 (4.0.2): do not edit by hand
\name{writeSimulatedInbixDataset}
\alias{writeSimulatedInbixDataset}
\title{Write simulated differential expression data matrix as inbix format}
\usage{
writeSimulatedInbixDataset(D, Dfileprefix)
}
\arguments{
\item{D}{is a simulated differential coexpression data matrix}

\item{Dfileprefix}{is an output data file prefix}
}
\description{
\code{writeSimulatedInbixDataset}
}

