% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inbixCppInterface.R
\name{writeSimulatedInbixDataset}
\alias{writeSimulatedInbixDataset}
\title{Write simulated differential expression data matrix as inbix format.}
\usage{
writeSimulatedInbixDataset(D, Dfileprefix)
}
\arguments{
\item{D}{Data frame simulated differential coexpression.}

\item{Dfileprefix}{String output data file prefix.}
}
\description{
\code{writeSimulatedInbixDataset}
}

